+++ src/briefy/leica/views/dashboard/customer.py
@@ -117,7 +117,7 @@ class DashboardCustomerDeliveredOrderService(SQLQueryService):
 
     (SELECT i.id, i.state, i.title, o.accept_date, o.project_id
     FROM items as i JOIN orders as o on i.id = o.id
-    WHERE i.type = '{type}' AND
+    WHERE o.current_type = '{type}' AND
     i.state IN ('delivered', 'accepted', 'in_qa', 'refused')
     ) as orders JOIN
 
@@ -220,7 +220,7 @@ class DashboardCustomerAllLeadsService(SQLQueryService):
     (SELECT i.id, i.state, i.title, o.accept_date, o.project_id, o.current_type
     FROM items as i JOIN orders as o on i.id = o.id
     JOIN leadorders as l on l.id = o.id
-    WHERE i.type = '{type}' AND
+    WHERE o.current_type = '{type}' AND
     i.state IN ('new', 'received', 'assigned', 'scheduled', 'cancelled',
     'delivered', 'accepted', 'in_qa', 'refused', 'perm_refused')
     ) as orders JOIN