+++ HISTORY.rst
@@ -98,6 +98,7 @@ History
     * Fix all to_dict methods so now they can be cached without pickle errors (rudaporto).
     * Card #582: fix assignment slug creation based on the order slug (rudaporto).
     * Card #583: fix leadorder confirm transition to create a new assignment only if there is no one (rudaporto).
+    * Card #590: fixed total of orders / leads in the dashboards for customer, bizdev, pm and finance (rudaporto).
 
 2.1.36 (2017-08-06)
 -------------------